pyproject.toml
[project]
name = "docling-surya"
version = "0.1.0"
description = "Docling plugin for Surya OCR"
readme = "README.md"
authors = [
  { name = "Harry Ho", email = "harrykhh@gmail.com" }
]
keywords = ["docling", "ocr", "layout", "plugin"]
repository = "https://github.com/harrykhh/docling_surya"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.60.0",
    "surya-ocr>=0.17.0",
]

[project.entry-points.docling]
"surya-ocr" = "docling_surya.plugin"

[dependency-groups]
dev = [
    "pytest>=7.0",
    "ruff>=0.5.0",
    "reportlab>=4.4.4",
    "docling-surya",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
docling-surya = { workspace = true }

[tool.ruff]
target-version = "py312"
line-length = 88
respect-gitignore = true

[tool.ruff.format]
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = [
  "B",  # flake8-bugbear
  "C",  # flake8-comprehensions
  "C9", # mccabe
  "D",  # flake8-docstrings
  "E",   # pycodestyle errors (default)
  "F",   # pyflakes (default)
  "I",   # isort
  "PD",  # pandas-vet
  "PIE", # pie
  "PTH", # pathlib
  "Q", # flake8-quotes
  "RET", # return
  "RUF", # Enable all ruff-specific checks
  "SIM", # simplify
  "S307", # eval
  "T20",  # (disallow print statements) keep debugging statements out of the codebase
  "W",     # pycodestyle warnings
  "ASYNC", # async
  "UP",    # pyupgrade
]

ignore = [
  "C408",   # Unnecessary `dict()` call (rewrite as a literal)
  "E501",   # Line too long, handled by ruff formatter
  "D107",   # "Missing docstring in __init__",
  "F401",   # imported but unused; consider using `importlib.util.find_spec` to test for "
  "F811",   # "redefinition of the same function"
  "PL",     # Pylint
  "RUF012", # Mutable Class Attributes
  "UP006",  # List vs list, etc
  "UP007",  # Option and Union
  "UP035",  # `typing.Set` is deprecated, use `set` instead"
]

[tool.ruff.lint.pep8-naming]
classmethod-decorators = [
  # Allow Pydantic's `@validator` decorator to trigger class method treatment.
  "pydantic.validator",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]
"tests/*.py" = ["ASYNC"]         # Disable ASYNC check for tests

[tool.ruff.lint.mccabe]
max-complexity = 30

[tool.ruff.lint.isort]
combine-as-imports = true